backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-groq
langgraph
serpapi
requests
sarvamai
appwrite
google-search-results